Safari-609.1.11.1/LayoutTests/crypto/subtle/aes-ctr-generate-export-key-jwk-length-256.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test exporting a 256 bits AES-CTR key with JWK format");

jsTestIsAsync = true;

var extractable = true;

debug("Generating a key...");
crypto.subtle.generateKey({name: "aes-ctr", length: 256}, extractable, ["encrypt", "decrypt", "wrapKey", "unwrapKey"]).then(function(key) {
    debug("Exporting a key...");
    return crypto.subtle.exportKey("jwk", key);
}).then(function(result) {
    key = result;

    shouldBe("key.kty", "'oct'");
    shouldBe("key.key_ops", "['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']");
    shouldBe("key.alg", "'A256CTR'");
    shouldBe("key.ext", "true");
    shouldBe("Base64URL.parse(key.k).byteLength", "32");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
